move and tweak the mw-indicators
mediawiki has indicators that are used to put icons at the
top of a page. I dont think any wikipages use this, but the
way the mw-indicator div was positioned was causing issues
with content at the top of the page that was floated.

This commit moves the mw-indicator div to the page header,
and tweaks the padding a little bit to not move things around.

fixes #2 #21