use User->isRegistered() rather than isLoggedIn()
in Mediawiki 1.36, the User->isLoggedIn() method was deprecated in
favour of the method it wrapped User->isRegistered(). This method was
subsequently removed in Mediawiki 1.38. In the Fedora mediawiki theme,
we used the isLoggedIn() in the main template to show or hide the login
button and other elements for the logged in user. Consequently, the Fedora
mediawiki theme was no longer working when trying to use on mediawiki later
than 1.38. In this update, we now use the User->isRegistered() method
which resolves this issue.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>